requirements.txt
Pillow
gradio_client
fastapi
fastapi[uvicorn]
boto3
requests